+++ CHANGELOG.md
@@ -51,6 +51,10 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
   Removing this automatic behavior was necessary to fix
   [#945](https://github.com/nsidc/earthaccess/issues/945).
+  
+ - Removed `keyword` and `type` kwargs from `search_services` documentation to be consistent with `search_for_services` documentation.
+  ([#1093](https://github.com/nsidc/earthaccess/issues/1093))
+  ([@battistowx](https://github.com/battistowx))
 
 ### Fixed